chore(jans-auth-server): move ClientAuthorizations to jans-auth-persistence-model for reusing by jans-config-api #3798
